requirements.txt
python-dotenv
azure-search-documents==11.6.0b4
azure-storage-blob
azure-identity
openai
Halo
backoff
streamlit
tiktoken
langchain